PRC-256 : API - PATCH for contact relationship - send domain event on update of contact relationship - fix merge conflicts, move face to its dedicated folder
