+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/resource/ContactControllerTest.kt
@@ -14,7 +14,7 @@ import org.springframework.data.domain.Page
 import org.springframework.data.domain.PageImpl
 import org.springframework.data.domain.PageRequest
 import org.springframework.http.HttpStatus
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.facade.ContactPatchFacade
+import uk.gov.justice.digital.hmpps.hmppscontactsapi.facade.ContactFacade
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.helpers.createContactAddressDetails
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.helpers.createContactEmailDetails
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.helpers.createContactIdentityDetails
@@ -29,7 +29,6 @@ import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.patch.PatchCo
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.patch.PatchContactResponse
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response.ContactDetails
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response.ContactSearchResultItem
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.service.patch.ContactFacade
 import java.net.URI
 import java.time.LocalDate
 import java.time.LocalDateTime
@@ -289,16 +288,16 @@ class ContactControllerTest {
 
     @Test
     fun `should update a contact relationship successfully`() {
-      doNothing().whenever(contactPatchFacade).patchRelationship(id, prisonerContactId, request)
+      doNothing().whenever(contactFacade).patchRelationship(id, prisonerContactId, request)
 
       controller.patchContactRelationship(id, prisonerContactId, request)
 
-      verify(contactPatchFacade).patchRelationship(id, prisonerContactId, request)
+      verify(contactFacade).patchRelationship(id, prisonerContactId, request)
     }
 
     @Test
     fun `should propagate exceptions patching a contact relationship`() {
-      whenever(contactPatchFacade.patchRelationship(id, prisonerContactId, request)).thenThrow(RuntimeException("Bang!"))
+      whenever(contactFacade.patchRelationship(id, prisonerContactId, request)).thenThrow(RuntimeException("Bang!"))
 
       assertThrows<RuntimeException>("Bang!") {
         controller.patchContactRelationship(id, prisonerContactId, request)
